test: 로또 번호 검증 테스트 분리
