+++ src/test/java/lotto/model/LottoTest.java
@@ -23,29 +23,6 @@ class LottoTest {
         );
     }
 
-    @DisplayName("로또 번호의 개수가 6개가 넘어가면 예외가 발생한다.")
-    @Test
-    void createLottoByOverSize() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 6, 7)))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("로또 번호의 갯수 6개를 만족해야합니다.");
-    }
-
-    @DisplayName("로또 번호에 중복된 숫자가 있으면 예외가 발생한다.")
-    @Test
-    void createLottoByDuplicatedNumber() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 5)))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("숫자는 중복될 수 없습니다.");
-    }
-
-    @DisplayName("로또 번호의 숫자가 1~45의 범위를 만족하지 않을시 예외가 발생한다.")
-    @Test
-    void createLottoByImproperNumber() {
-        assertThatThrownBy(() -> new Lotto(List.of(1, 2, 3, 4, 5, 47)))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("숫자는 1에서 45사이여야합니다.");
-    }
 
     @DisplayName("당첨번호 : [1,3,7,8,9,11] ,보너스 번호 : 18 인경우 등수 판별")
     @ParameterizedTest(name = "로또번호 :{0}, 예상등수 : {1}")
